type-safe object.entries + moved App
