+++ src/components/App.tsx
@@ -4,8 +4,10 @@ import Metric from "components/Metric";
 import Plot from "components/Plot";
 import { GlobalContext } from "context";
 import { clx } from "helpers";
+import * as data from "parsedData";
 import React, { useContext } from "react";
 
+console.log({ ...data });
 
 const App = () => {
   const global = useContext(GlobalContext);
