+++ src/index.tsx
@@ -2,7 +2,7 @@ import { GlobalProvider } from "context";
 import React from "react";
 import ReactDOM from "react-dom";
 import "styles/index.css";
-import App from "./App";
+import App from "./components/App";
 
 ReactDOM.render(
   <GlobalProvider>
